+++ scripts/validation/fast_hnsw_validation.py
@@ -51,7 +51,7 @@ class FastHNSWValidator:
                 title VARCHAR(1000),
                 text_content LONGVARCHAR,
                 embedding_varchar VARCHAR(50000),
-                embedding_vector VECTOR(DOUBLE, 384),
+                embedding_vector VECTOR(FLOAT, 384),
                 PRIMARY KEY (doc_id)
             )
             """
